+++ Craftsman/Builders/Seeders/SeederBuilder.cs
@@ -15,13 +15,13 @@
 
     public class SeederBuilder
     {
-        public static void AddSeeders(string solutionDirectory, List<Entity> entities, string dbContextName, string projectBaseName = "")
+        public static void AddSeeders(string solutionDirectory, List<Entity> entities, string dbContextName, string projectBaseName)
         {
             try
             {
                 foreach (var entity in entities)
                 {
-                    var classPath = ClassPathHelper.SeederClassPath(solutionDirectory, $"{Utilities.GetSeederName(entity)}.cs");
+                    var classPath = ClassPathHelper.SeederClassPath(solutionDirectory, $"{Utilities.GetSeederName(entity)}.cs", projectBaseName);
 
                     if (!Directory.Exists(classPath.ClassDirectory))
                         Directory.CreateDirectory(classPath.ClassDirectory);