+++ Craftsman/Builders/Seeders/SeederModifier.cs
@@ -21,7 +21,7 @@
             {
                 foreach (var entity in entities)
                 {
-                    var classPath = ClassPathHelper.SeederClassPath(solutionDirectory, $"{Utilities.GetSeederName(entity)}.cs");
+                    var classPath = ClassPathHelper.SeederClassPath(solutionDirectory, $"{Utilities.GetSeederName(entity)}.cs", projectBaseName);
 
                     if (!Directory.Exists(classPath.ClassDirectory))
                         Directory.CreateDirectory(classPath.ClassDirectory);