+++ CHANGELOG.md
@@ -9,6 +9,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add Entity is Broken where I say `EntityBrokenHere`
   - need to update addentity command to use an add entity template that includes a `ProjectBaseName` that must match the base of the api and the core
+- consolidated projects
+- projects will now have a prefix of the solution name before each project type. for example, the api project with a solution name of `ordering` is `ordering.webapi`
 
 ## [0.8.2] - 2021-02-25